added kyc layout logic
