+++ app/dashboard/business-profile-setup/business-profile.module.css
@@ -12,16 +12,6 @@
   width: 33%;
 }
 
-.lhs_levels_ctn .lhs_kyc{
-  margin-bottom: var(--screen-padding-top);
-}
-
-@media screen and (max-width: 1000px) {
-  .lhs_levels_ctn {
-    display: flex;
-  }
-}
-
 .pages {
   padding: 0 var(--outer-component-mg);
   display: flex;
