+++ app/dashboard/kyc/corporate/address/page.js
@@ -1,14 +1,14 @@
-import AddressCop from "@/components/KYC/corporate/AddressCop";
+import Address from "@/components/KYC/corporate/Address";
 import styles from "../../kyc.module.css";
 
 export const metadata = () => {
-  return {
-    title: "Address - KYC | Passpoint Go",
-    description: "",
-  };
+	return {
+		title: "Address - KYC | Passpoint Go",
+		description: "",
+	};
 };
-const Address = () => {
-  return <AddressCop styles={styles} />;
+const AddressPage = () => {
+	return <Address styles={styles} />;
 };
 
-export default Address;
+export default AddressPage;
