+++ app/dashboard/kyc/corporate/verify-bvn/page.js
@@ -1,0 +1,11 @@
+import VerifyBVN from '@/components/KYC/corporate/VerifyBVN'
+
+const VerifyBVNPage = () => {
+	return (
+		<>
+			<VerifyBVN />
+		</>
+	)
+}
+
+export default VerifyBVNPage
