+++ assets/styles/globals.css
@@ -127,6 +127,7 @@
 
   --btn-fs: 16px;
   --btn-height: 48px;
+  --btn-height-sd: 48px;
   --btn-height-md: 40px;
   --btn-br: 12px;
   --btn-br-sd: 12px;
@@ -214,6 +215,11 @@ button * {
 a:disabled,
 button:disabled {
   opacity: 0.5;
+  cursor: not-allowed;
+}
+
+input:disabled {
+  cursor: not-allowed;
 }
 
 .skeleton {
@@ -340,7 +346,6 @@ li {
 .primary_btn,
 button.primary,
 .primary_link {
-  width: 100%;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -351,6 +356,14 @@ button.primary,
   border: none;
   background-color: var(--cta-bg);
   color: #fff;
+}
+
+.primary.half {
+  width: 50%;
+}
+
+button.half {
+  width: 50%;
 }
 
 .primary_btn:hover,
@@ -509,6 +522,10 @@ button.tertiary:hover,
   justify-content: center;
   gap: 8px;
   box-shadow: var(--btn-secondary-shadow);
+}
+
+.back_btn.sd {
+  height: var(--btn-height-sd);
 }
 
 .back_btn:hover {
