+++ components/Custom/Select/index.module.css
@@ -39,6 +39,11 @@
   box-shadow: 0px 0px 0px 3px var(--input-error-box-shadow) !important;
 }
 
+.custom_select button:disabled {
+  background-color: var(--grey-5);
+  border: 1px solid var(--input-border-disabled) !important;
+}
+
 .custom_select button svg {
   transition: 0.2s;
 }
