--- components/KYC/SetupLHS/setup-lhs.css
@@ -1,5 +1,0 @@
-@media screen and (max-width: 1000px) {
-  .levels_ctn {
-    display: flex;
-  }
-}
